fix: address PR #4082 second-round review (Tier 1 + Tier 2)
Multi-model review (4 internal + Codex CLI) surfaced several findings.
Tier 1 (blocking) and Tier 2 (high-risk) addressed here; Tier 3 stays
in #4103.

Production-safety:
- systemd: ProtectSystem=strict blocked the sudo'd child from writing
  /usr/local/bin/freenet — Phase 2 would have failed on first real
  update. Switched to ProtectSystem=off and RestrictNamespaces=false
  with comments explaining why the privilege boundary is sudoers, not
  the systemd sandbox. (Codex P1)
- gateway-update.yml: `max-parallel` was at job level instead of
  `strategy.max-parallel` — workflow was invalid YAML, couldn't be
  dispatched. (Codex P2)
- server.rs: mutex acquire wrapped in tokio::time::timeout(5s) so a
  stuck spawn can't deadlock all future /update requests indefinitely.
  Returns 503 on timeout. (Skeptical H1)
- sudoers.freenet-release-agent: comment claimed `*` matches "a single
  non-whitespace argument" — false per man 5 sudoers, `*` matches any
  characters including whitespace. Fixed comment to explain the safety
  comes from two independent semver regex checks (agent semver::parse,
  then bash regex in gateway-auto-update.sh), not sudoers alone.
  (Code-first + Skeptical H2)
- gateway-auto-update.sh: added strict semver regex validation of
  --target-version after the `${2#v}` strip — defense in depth against
  the sudoers wildcard. (Code-first)
- version.rs: VersionCache keyed on mtime alone would return stale
  data forever if the update script used `cp -p` (mtime preserved on
  rename). Added size to the cache key + a regression test that
  preserves mtime and changes size. (Skeptical M5, rule-review WARNING)
- docs/release-agent-deployment.md: claimed "no in-memory cache" for
  the HMAC secret, but main.rs reads it once at startup. Fixed the
  rotation procedure to require `systemctl restart`. (Skeptical L4)
- config.rs: clock_skew_tolerance_seconds changed i64 → u32 so a
  misconfigured negative value can't accidentally reject every
  request. (Skeptical L3)

Test coverage:
- spawn_failure_does_not_consume_rate_limit_window — pins the security
  contract that an immediate sudo failure (exit 1 within 1s probe) must
  NOT consume the rate-limit window. (Testing high-risk)
- sudoers_argv_matches_allowlist — fake-sudo records the agent's
  argv and asserts it matches the sudoers entry's prefix. Future
  refactors that reorder flags will fail this test before reaching
  production. (Testing high-risk)
- oversized_body_rejected — pins the 4KiB body limit. (Code-first)
- cache_invalidates_when_binary_changes — regression test for the
  size-keyed cache fix above.
- Added a build_live() variant on the test Harness that injects a
  fake-sudo via the new Updater.sudo_command field, so the spawn
  pipeline can be exercised without real root.

Docs:
- AGENTS.md: added crates/release-agent/ to Repository Structure.
- scripts/release.sh: future-migration comment on trigger_gateway_updates
  pointing at .github/workflows/gateway-update.yml and #4073.
- gateway-auto-update.sh: usage block documents --target-version.

Workflow:
- gateway-update.yml: added comment documenting the split-version risk
  of fail-fast + max-parallel:1 on partial-failure rollout.

Test totals: 37 (23 unit + 14 integration). cargo fmt + clippy
--all-targets -D warnings clean.

+++ crates/release-agent/src/auth.rs
@@ -38,7 +38,10 @@ pub fn verify_signature(secret: &[u8], body: &[u8], header_value: &str) -> Resul
 /// Reject requests whose `issued_at` is non-positive or outside ±tolerance
 /// of the agent's wall clock. Uses checked arithmetic so an attacker who
 /// produces a valid signature cannot panic the agent with `i64::MIN`.
-pub fn check_clock_skew(issued_at: i64, tolerance_seconds: i64) -> Result<(), AuthError> {
+///
+/// `tolerance_seconds` is unsigned so a misconfigured negative value
+/// can't accidentally reject every request.
+pub fn check_clock_skew(issued_at: i64, tolerance_seconds: u32) -> Result<(), AuthError> {
     if issued_at <= 0 {
         return Err(AuthError::NonPositiveIssuedAt);
     }
@@ -50,7 +53,7 @@ pub fn check_clock_skew(issued_at: i64, tolerance_seconds: i64) -> Result<(), Au
         .checked_sub(issued_at)
         .and_then(i64::checked_abs)
         .unwrap_or(i64::MAX);
-    if delta > tolerance_seconds {
+    if delta > tolerance_seconds as i64 {
         return Err(AuthError::ClockSkew(delta));
     }
     Ok(())
